pyproject.toml
[tool.poetry]
name = "dakota"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gunicorn = "^23.0.0"
fastapi = "^0.115.4"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
openpyxl = "^3.1.5"
reportlab = "^4.2.5"
python-multipart = "^0.0.17"
pydantic-settings = "^2.6.1"
pydantic = {extras = ["email"], version = "^2.9.2"}


[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = "^0.32.0"}
faker = "^30.8.2"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
